chore: move add members and new project
