+++ webui/react/src/pages/WorkspaceDetails/WorkspaceDetailsHeader.module.scss
@@ -21,13 +21,6 @@
   .dropdown {
     font-size: 12px;
   }
-  .headerButton {
-    display: inline-flex;
-
-    button {
-      margin-right: 10px;
-    }
-  }
 
   @media only screen and (max-width: $breakpoint-tablet) {
     display: block;
