+++ webui/react/src/pages/WorkspaceDetails/WorkspaceDetailsHeader.tsx
@@ -4,11 +4,7 @@ import React, { useCallback } from 'react';
 
 import DynamicIcon from 'components/DynamicIcon';
 import InlineEditor from 'components/InlineEditor';
-import Button from 'components/kit/Button';
 import Tooltip from 'components/kit/Tooltip';
-import useFeature from 'hooks/useFeature';
-import useModalProjectCreate from 'hooks/useModal/Project/useModalProjectCreate';
-import useModalWorkspaceAddMember from 'hooks/useModal/Workspace/useModalWorkspaceAddMember';
 import usePermissions from 'hooks/usePermissions';
 import WorkspaceActionDropdown from 'pages/WorkspaceList/WorkspaceActionDropdown';
 import { patchWorkspace } from 'services/api';
@@ -27,37 +23,8 @@ interface Props {
   workspace: Workspace;
 }
 
-const WorkspaceDetailsHeader: React.FC<Props> = ({
-  addableUsersAndGroups,
-  rolesAssignableToScope,
-  workspace,
-  fetchWorkspace,
-}: Props) => {
-  const { canAssignRoles } = usePermissions();
-
-  const { contextHolder, modalOpen: openProjectCreate } = useModalProjectCreate({
-    workspaceId: workspace.id,
-  });
-
-  const { contextHolder: workspaceAddMemberContextHolder, modalOpen: openWorkspaceAddMember } =
-    useModalWorkspaceAddMember({
-      addableUsersAndGroups,
-      onClose: fetchWorkspace,
-      rolesAssignableToScope,
-      workspaceId: workspace.id,
-    });
-
-  const rbacEnabled = useFeature().isOn('rbac');
-
-  const { canCreateProject, canModifyWorkspace } = usePermissions();
-
-  const handleProjectCreateClick = useCallback(() => {
-    openProjectCreate();
-  }, [openProjectCreate]);
-
-  const handleAddMembersClick = useCallback(() => {
-    openWorkspaceAddMember();
-  }, [openWorkspaceAddMember]);
+const WorkspaceDetailsHeader: React.FC<Props> = ({ workspace, fetchWorkspace }: Props) => {
+  const { canModifyWorkspace } = usePermissions();
 
   const handleNameChange = useCallback(
     async (name: string) => {
@@ -114,19 +81,6 @@ const WorkspaceDetailsHeader: React.FC<Props> = ({
           </WorkspaceActionDropdown>
         )}
       </Space>
-      <div className={css.headerButton}>
-        {rbacEnabled &&
-          canAssignRoles({ workspace }) &&
-          !workspace.immutable &&
-          !workspace.archived && <Button onClick={handleAddMembersClick}> Add Members</Button>}
-        {!workspace.immutable &&
-          !workspace.archived &&
-          canCreateProject({ workspace: workspace }) && (
-            <Button onClick={handleProjectCreateClick}>New Project</Button>
-          )}
-      </div>
-      {workspaceAddMemberContextHolder}
-      {contextHolder}
     </div>
   );
 };
