+++ webui/react/src/pages/WorkspaceDetails/WorkspaceMembers.module.scss
@@ -1,0 +1,6 @@
+.headerButton {
+  align-items: flex-end;
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 16px;
+}
