+++ webui/react/src/pages/WorkspaceDetails/WorkspaceMembers.tsx
@@ -10,6 +10,8 @@ import SkeletonTable from 'components/Table/SkeletonTable';
 import { getFullPaginationConfig } from 'components/Table/Table';
 import TableFilterSearch from 'components/Table/TableFilterSearch';
 import UserBadge from 'components/UserBadge';
+import useFeature from 'hooks/useFeature';
+import useModalWorkspaceAddMember from 'hooks/useModal/Workspace/useModalWorkspaceAddMember';
 import useModalWorkspaceRemoveMember from 'hooks/useModal/Workspace/useModalWorkspaceRemoveMember';
 import usePermissions from 'hooks/usePermissions';
 import { UpdateSettings, useSettings } from 'hooks/useSettings';
@@ -21,12 +23,14 @@ import { User, UserOrGroup, Workspace } from 'types';
 import { getAssignedRole, getIdFromUserOrGroup, getName, isUser } from 'utils/user';
 
 import RoleRenderer from './RoleRenderer';
+import css from './WorkspaceMembers.module.scss';
 import settingsConfig, {
   DEFAULT_COLUMN_WIDTHS,
   WorkspaceMembersSettings,
 } from './WorkspaceMembers.settings';
 
 interface Props {
+  addableUsersAndGroups: UserOrGroup[];
   assignments: V1RoleWithAssignments[];
   fetchMembers: () => void;
   groupsAssignedDirectly: V1Group[];
@@ -98,6 +102,7 @@ const GroupOrMemberActionDropdown: React.FC<GroupOrMemberActionDropdownProps> = 
 };
 
 const WorkspaceMembers: React.FC<Props> = ({
+  addableUsersAndGroups,
   assignments,
   onFilterUpdate,
   usersAssignedDirectly,
@@ -113,9 +118,23 @@ const WorkspaceMembers: React.FC<Props> = ({
 
   const usersAndGroups: UserOrGroup[] = [...usersAssignedDirectly, ...groupsAssignedDirectly];
 
+  const { contextHolder: workspaceAddMemberContextHolder, modalOpen: openWorkspaceAddMember } =
+    useModalWorkspaceAddMember({
+      addableUsersAndGroups,
+      onClose: fetchMembers,
+      rolesAssignableToScope,
+      workspaceId: workspace.id,
+    });
+
+  const rbacEnabled = useFeature().isOn('rbac');
+
   useEffect(() => {
     onFilterUpdate(settings.name);
   }, [onFilterUpdate, settings.name]);
+
+  const handleAddMembersClick = useCallback(() => {
+    openWorkspaceAddMember();
+  }, [openWorkspaceAddMember]);
 
   const handleNameSearchApply = useCallback(
     (newSearch: string) => {
@@ -223,6 +242,12 @@ const WorkspaceMembers: React.FC<Props> = ({
 
   return (
     <>
+      <div className={css.headerButton}>
+        {rbacEnabled &&
+          canAssignRoles({ workspace }) &&
+          !workspace.immutable &&
+          !workspace.archived && <Button onClick={handleAddMembersClick}> Add Members</Button>}
+      </div>
       {settings ? (
         <InteractiveTable
           columns={columns}
@@ -244,6 +269,7 @@ const WorkspaceMembers: React.FC<Props> = ({
       ) : (
         <SkeletonTable columns={columns.length} />
       )}
+      {workspaceAddMemberContextHolder}
     </>
   );
 };
